Fixes an extract tags bug and adds tag dict shift
There was a minor bug when attempting to shift tags saved by the
extract_text_tags function, because the string type returned by the
tag_ranges tk method is not a Python string but a special tk string
object that mimics a Python string. The string methods used in the
tag shifting functions would not work with these tk strings.

The new shift_tag_dict function allows entire dictionaries of tags
to be shifted by the shift_tag_range function.