spelldb can get the spell list and spells by id
Adds methods to the SpellDataBase class to get the full list of spells
in the database with their corresponding row ids. Another method gets
spells from the database using their ID number.

Also changes the definition of the TagRange for storing text formatting
tags. When a TagDict is stored using the json dumps, then retrieved by
loads, the tag range tuples are converted to lists. Since it isn't
necessary that the tag ranges be expressed as tuples, the definition was
changed to use lists.